refactor: capitalize names of models
