+++ middlewares/passport.js
@@ -1,71 +1,70 @@
 const config = require("../config/config");
 const passport = require("passport");
 const { ExtractJwt, Strategy } = require("passport-jwt");
-const User = require("../models/users");
+const User = require("../models/User");
 
 const initialize = () => {
-    passport.use("jwt", getStrategy());
-    return passport.initialize();
+  passport.use("jwt", getStrategy());
+  return passport.initialize();
 };
 
 const authenticate = (req, res, next) => {
-    try {
-        return passport.authenticate(
-            "jwt",
-            { session: false },
-            (err, user, info) => {
-                if (err) {
-                    return next(err);
-                }
+  try {
+    return passport.authenticate(
+      "jwt",
+      { session: false },
+      (err, user, info) => {
+        if (err) {
+          return next(err);
+        }
 
-                if (!user) {
-                    if (info.name === "TokenExpiredError") {
-                        return res.status(401).json({
-                            message:
-                                "Your token has expired. Please generate a new one",
-                        });
-                    } else {
-                        return res.status(401).json({ message: info.message });
-                    }
-                }
+        if (!user) {
+          if (info.name === "TokenExpiredError") {
+            return res.status(401).json({
+              message: "Your token has expired. Please generate a new one",
+            });
+          } else {
+            return res.status(401).json({ message: info.message });
+          }
+        }
 
-                req.user = user;
-                return next();
-            }
-        )(req, res, next);
-    } catch (error) {
-        console.log(error);
-        res.status(500).json({
-            error: true,
-            message: error.message || "Произошла какая-та ошибка",
-        });
-    }
+        req.user = user;
+        return next();
+      }
+    )(req, res, next);
+  } catch (error) {
+    console.log(error);
+    res.status(500).json({
+      error: true,
+      message: error.message || "Произошла какая-та ошибка",
+    });
+  }
 };
 
 const getStrategy = () => {
-    const params = {
-      secretOrKey: config,
-      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-    };
+  const params = {
+    secretOrKey: config,
+    jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+  };
 
-    return new Strategy(params, async (payload, done) => {
-        await User.findById(payload.id, (err, user) => {
-            if (err) {
-                return done(err);
-            }
+  return new Strategy(params, async (payload, done) => {
+    await User.findById(payload.id, (err, user) => {
+      if (err) {
+        return done(err);
+      }
 
-            if (!user) {
-                return done(null, false, {
-                    message: "The user in the token was not found",
-                });
-            }
-
-            return done(null, user);
+      if (!user) {
+        return done(null, false, {
+          message: "The user in the token was not found",
         });
+      }
+
+      return done(null, user);
     });
+  });
 };
 
 module.exports = {
-    initialize,
-    authenticate,
+  initialize,
+  authenticate,
 };
